Simplify Breadcrumbs by making immediate children a BreadcrumbItem automatically
