+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "fictoan-react",
-    "version": "1.9.3",
+    "version": "1.9.4-alpha.1",
     "private": false,
     "description": "A full-featured, designer-friendly, yet performant framework with plain-English props and focus on rapid iteration.",
     "repository": {
