--- src/components/Breadcrumbs/Breadcrumbs/breadcrumbs.css
@@ -1,11 +1,0 @@
-[data-breadcrumbs-wrapper] {
-    margin      : 0; /* Reset */
-    padding     : 0; /* Reset */
-    display     : flex;
-    flex-wrap   : wrap;
-    align-items : flex-start;
-
-    &:not([class*="bg-"]) {
-        background-color : var(--breadcrumbs-wrapper-bg);
-    }
-}
