+++ src/components/Breadcrumbs/index.tsx
@@ -1,3 +1,1 @@
-export { Breadcrumbs, type BreadcrumbsProps } from "../Breadcrumbs/Breadcrumbs/Breadcrumbs";
-
-export { BreadcrumbItem, type BreadcrumbItemProps } from "./BreadcrumbItem/BreadcrumbItem";
+export { Breadcrumbs, type BreadcrumbsProps } from "./Breadcrumbs";
